refactor: move store to its own directory
